// lib/api/generated/models/ReservationResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ClassHub API
 * ClassHub Backend API 문서
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 예약 응답
 * @export
 * @interface ReservationResponse
 */
export interface ReservationResponse {
    /**
     * 예약 ID
     * @type {number}
     * @memberof ReservationResponse
     */
    reservationId?: number;
    /**
     * 학생 ID
     * @type {number}
     * @memberof ReservationResponse
     */
    studentId?: number;
    /**
     * 신청자 이름
     * @type {string}
     * @memberof ReservationResponse
     */
    applicantName?: string;
    /**
     * 연락처
     * @type {string}
     * @memberof ReservationResponse
     */
    phoneNumber?: string;
    /**
     * 신청 일시
     * @type {Date}
     * @memberof ReservationResponse
     */
    appliedAt?: Date;
    /**
     * 예약 상태
     * @type {string}
     * @memberof ReservationResponse
     */
    reservationStatus?: string;
    /**
     * D-3 알림 발송 여부
     * @type {boolean}
     * @memberof ReservationResponse
     */
    sentD3Notification?: boolean;
    /**
     * D-1 알림 발송 여부
     * @type {boolean}
     * @memberof ReservationResponse
     */
    sentD1Notification?: boolean;
}

/**
 * Check if a given object implements the ReservationResponse interface.
 */
export function instanceOfReservationResponse(value: object): value is ReservationResponse {
    return true;
}

export function ReservationResponseFromJSON(json: any): ReservationResponse {
    return ReservationResponseFromJSONTyped(json, false);
}

export function ReservationResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ReservationResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'reservationId': json['reservationId'] == null ? undefined : json['reservationId'],
        'studentId': json['studentId'] == null ? undefined : json['studentId'],
        'applicantName': json['applicantName'] == null ? undefined : json['applicantName'],
        'phoneNumber': json['phoneNumber'] == null ? undefined : json['phoneNumber'],
        'appliedAt': json['appliedAt'] == null ? undefined : (new Date(json['appliedAt'])),
        'reservationStatus': json['reservationStatus'] == null ? undefined : json['reservationStatus'],
        'sentD3Notification': json['sentD3Notification'] == null ? undefined : json['sentD3Notification'],
        'sentD1Notification': json['sentD1Notification'] == null ? undefined : json['sentD1Notification'],
    };
}

export function ReservationResponseToJSON(json: any): ReservationResponse {
    return ReservationResponseToJSONTyped(json, false);
}

export function ReservationResponseToJSONTyped(value?: ReservationResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'reservationId': value['reservationId'],
        'studentId': value['studentId'],
        'applicantName': value['applicantName'],
        'phoneNumber': value['phoneNumber'],
        'appliedAt': value['appliedAt'] == null ? value['appliedAt'] : value['appliedAt'].toISOString(),
        'reservationStatus': value['reservationStatus'],
        'sentD3Notification': value['sentD3Notification'],
        'sentD1Notification': value['sentD1Notification'],
    };
}